Documentation cleanup, removal of (some) unfinished pages
A lot of pages that describe things like the upload process or FAQs are
already present on the documentation in the actual site. Adding them
here as well seems unnecessary, though we should definitely link to the
help pages.

The idea of the RTD site should be about contributing to the project,
not so much about using it.