Fix issues with non-valued measurements
Ideally, it would be nice if we could avoid creating records in the
first place, but SQLAlchemy throws a bunch of warnings if we skip adding
the measurement contexts to the session.

The reason for this is that the records' relationships (study,
bioreplicate, etc), trigger changes in the in-memory representations of
said study, bioreplicate, etc, and there's no way that I can see of
"undoing" these changes.